changed external structure
