+++ index.spec.ts
@@ -1,13 +1,13 @@
-import * as service from "./index"
+import * as ch3d1 from "./index"
 
 describe("service-ch3d1", () => {
 	it("test accessibility", () => {
-		service.ch3d1.Eci.is({})
-		const pareqFunctions = [service.ch3d1.Pareq.decode, service.ch3d1.Pareq.encode, service.ch3d1.Pareq.is]
-		const paresFunctions = [service.ch3d1.Pares.decode, service.ch3d1.Pares.encode, service.ch3d1.Pares.is]
-		const cardFunctions = [service.ch3d1.api.Card.from, service.ch3d1.api.Card.to, service.ch3d1.api.Card.is]
-		service.ch3d1.api.Configuration.is({})
-		const checkFunctions = [service.ch3d1.api.check.Request.is, service.ch3d1.api.check.Response.is, service.ch3d1.api.check.Response.verify, service.ch3d1.api.check.post]
-		const enrolledFunctions = [service.ch3d1.api.enrolled.Request.is, service.ch3d1.api.enrolled.Response.is, service.ch3d1.api.enrolled.post]
+		ch3d1.Eci.is({})
+		const pareqFunctions = [ch3d1.Pareq.decode, ch3d1.Pareq.encode, ch3d1.Pareq.is]
+		const paresFunctions = [ch3d1.Pares.decode, ch3d1.Pares.encode, ch3d1.Pares.is]
+		const cardFunctions = [ch3d1.api.Card.from, ch3d1.api.Card.to, ch3d1.api.Card.is]
+		ch3d1.api.Configuration.is({})
+		const checkFunctions = [ch3d1.api.check.Request.is, ch3d1.api.check.Response.is, ch3d1.api.check.Response.verify, ch3d1.api.check.post]
+		const enrolledFunctions = [ch3d1.api.enrolled.Request.is, ch3d1.api.enrolled.Response.is, ch3d1.api.enrolled.post]
 	})
 })
